* Added '\''listen'\'' playback mode for the sampler in dspu::Sampler.
* Added support of build for MacOS.
* Updated module versions in dependencies.

+++ include/lsp-plug.in/dsp-units/util/MeterGraph.h
@@ -99,7 +99,7 @@ namespace lsp
                  */
                 inline void set_period(size_t period)
                 {
-                    nPeriod         = period;
+                    nPeriod         = uint32_t(period);
                 }
 
                 inline float period() const             { return nPeriod; }
