+++ src/main/dynamics/DynamicProcessor.cpp
@@ -1,6 +1,6 @@
 /*
- * Copyright (C) 2024 Linux Studio Plugins Project <https://lsp-plug.in/>
- *           (C) 2024 Vladimir Sadovnikov <sadko4u@gmail.com>
+ * Copyright (C) 2025 Linux Studio Plugins Project <https://lsp-plug.in/>
+ *           (C) 2025 Vladimir Sadovnikov <sadko4u@gmail.com>
  *
  * This file is part of lsp-dsp-units
  * Created on: 19 окт. 2016 г.
@@ -81,7 +81,7 @@ namespace lsp
         {
             if (sr == nSampleRate)
                 return;
-            nSampleRate = sr;
+            nSampleRate = uint32_t(sr);
             bUpdate     = true;
         }
