+++ src/main/dynamics/Gate.cpp
@@ -139,7 +139,7 @@ namespace lsp
         {
             if (sr == nSampleRate)
                 return;
-            nSampleRate = sr;
+            nSampleRate = uint32_t(sr);
             bUpdate     = true;
         }
